pyproject.toml
[tool.poetry]
name = "ekondo_expense_mgt"
version = "0.1.0"
description = "Ekondo Expense Management System"
authors = ["esunday5 <emmyblaq3@gmail.com>"]
license = "MIT"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
flask = "2.3.3"
flask-login = "0.6.3"
sqlalchemy = "2.0.36"
gunicorn = "21.2.0"
alembic = "1.14.0"
attrs = "24.2.0"
blinker = "1.9.0"
build = "1.2.2.post1"
cachecontrol = "0.14.1"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
crashtest = "0.4.1"
deprecated = "1.2.15"
distlib = "0.3.9"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.16.1"
flasgger = "0.9.7.1"
flask-cors = "5.0.0"
flask-dotenv = "0.1.2"
flask-limiter = "3.8.0"
flask-migrate = "4.0.7"
flask-sqlalchemy = "3.1.1"
flask-talisman = "1.1.0"
flask-wtf = "1.2.2"
greenlet = "3.1.1"
idna = "3.10"
importlib-resources = "6.4.5"
installer = "0.7.0"
itsdangerous = "2.2.0"
jaraco-classes = "3.4.0"
jinja2 = "3.1.4"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
keyring = "24.3.1"
limits = "3.13.0"
mako = "1.3.6"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
mdurl = "0.1.2"
mistune = "3.0.2"
more-itertools = "10.5.0"
msgpack = "1.1.0"
ordered-set = "4.1.0"
packaging = "24.2"
pexpect = "4.9.0"
pillow = "11.0.0"
pkginfo = "1.11.2"
platformdirs = "4.3.6"
poetry = "1.8.4"
poetry-core = "1.9.1"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pygments = "2.18.0"
pyproject-hooks = "1.2.0"
python-dotenv = "1.0.1"
pywin32-ctypes = "0.2.3"
pyyaml = "6.0.2"
rapidfuzz = "3.10.1"
referencing = "0.35.1"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.9.4"
rpds-py = "0.21.0"
shellingham = "1.5.4"
six = "1.16.0"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
virtualenv = "20.27.1"
werkzeug = "3.1.3"
wrapt = "1.17.0"
wtforms = "3.2.1"
waitress = "^3.0.2"
flask-caching = "^2.3.0"
pdf2image = "^1.17.0"
pymysql = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
